MapStore refactoring - step 3
Pulled up common methods from MapDB and BekerleyDB stores.
